+++ lib/src/all_models.g.dart
@@ -16,8 +16,10 @@ Serializers _$serializers = (new Serializers().toBuilder()
       ..add(Keyword.serializer)
       ..add(Rank.serializer)
       ..add(Unit.serializer)
+      ..add(UnitKey.serializer)
       ..add(UnitType.serializer)
       ..add(Upgrade.serializer)
+      ..add(UpgradeKey.serializer)
       ..add(UpgradeSlot.serializer)
       ..add(Weapon.serializer)
       ..addBuilderFactory(
@@ -50,8 +52,8 @@ Serializers _$serializers = (new Serializers().toBuilder()
               const [const FullType(Keyword), const FullType(String)]),
           () => new MapBuilder<Keyword, String>())
       ..addBuilderFactory(
-          const FullType(BuiltSet, const [const FullType(Upgrade)]),
-          () => new SetBuilder<Upgrade>()))
+          const FullType(BuiltSet, const [const FullType(UpgradeKey)]),
+          () => new SetBuilder<UpgradeKey>()))
     .build();
 
 // ignore_for_file: always_put_control_body_on_new_line,annotate_overrides,avoid_annotating_with_dynamic,avoid_as,avoid_catches_without_on_clauses,avoid_returning_this,lines_longer_than_80_chars,omit_local_variable_types,prefer_expression_function_bodies,sort_constructors_first,test_types_in_equals,unnecessary_const,unnecessary_new