+++ lib/src/model/unit.dart
@@ -75,73 +75,113 @@ abstract class Unit implements Built<Unit, UnitBuilder> {
   String get id;
 
   /// Name of the unit.
+  @BuiltValueField(compare: false)
   String get name;
 
   /// Whether this is a _unique_ unit (one per army).
-  @BuiltValueField(wireName: 'is_unique')
+  @BuiltValueField(compare: false, wireName: 'is_unique')
   bool get isUnique;
 
   /// An optional _sub-title_ for the unit.
   ///
   /// If omitted (`null`), then the unit does not have one.
-  @BuiltValueField(wireName: 'sub_title')
+  @BuiltValueField(compare: false, wireName: 'sub_title')
   @nullable
   String get subTitle;
 
   /// Faction the unit belongs to.
+  @BuiltValueField(compare: false)
   Faction get faction;
 
   /// Type of the unit.
+  @BuiltValueField(compare: false)
   UnitType get type;
 
   /// Amount of points the unit costs to bring into an army list.
+  @BuiltValueField(compare: false)
   int get points;
 
   /// Rank of the unit when building an army list.
+  @BuiltValueField(compare: false)
   Rank get rank;
 
   /// Number of miniatures that come with the unit.
+  @BuiltValueField(compare: false)
   int get miniatures;
 
   /// Amount of damage that can be inflicted on this unit.
+  @BuiltValueField(compare: false)
   int get wounds;
 
   /// Courage of the unit (for troopers).
   ///
   /// If omitted (`null`), this unit is immune to the effects of suppression.
+  @BuiltValueField(compare: false)
   @nullable
   int get courage;
 
   /// Resilience of the unit (for vehicles).
   ///
   /// If omitted (`null`), this unit is immune to the effects of vehicle damage.
+  @BuiltValueField(compare: false)
   @nullable
   int get resilience;
 
   /// Dice to be rolled to prevent damage.
+  @BuiltValueField(compare: false)
   DefenseDice get defense;
 
   /// How the _surge_ icon converts for _attacks_.
   ///
   /// If omitted (`null`), it does not convert.
-  @BuiltValueField(wireName: 'attack_surge')
+  @BuiltValueField(compare: false, wireName: 'attack_surge')
   @nullable
   AttackSurge get attackSurge;
 
   /// Whether the _surge_ icon converts for _defence_.
-  @BuiltValueField(wireName: 'has_defense_surge')
+  @BuiltValueField(compare: false, wireName: 'has_defense_surge')
   bool get hasDefenseSurge;
 
   /// Speed of the unit when moving.
+  @BuiltValueField(compare: false)
   int get speed;
 
   /// Available slots for upgrades for this unit.
-  @BuiltValueField(wireName: 'upgrades')
+  @BuiltValueField(compare: false, wireName: 'upgrades')
   BuiltMap<UpgradeSlot, int> get upgrades;
 
   /// Weapons included with this unit.
+  @BuiltValueField(compare: false)
   BuiltSet<Weapon> get weapons;
 
   /// Keywords on the unit.
+  @BuiltValueField(compare: false)
   BuiltMap<Keyword, String> get keywords;
+
+  /// Returns a [UnitKey] reference for this [Unit].
+  UnitKey toKey() => UnitKey(id: id, name: name);
+}
+
+/// Represents a reference to a [Unit] based on [id].
+abstract class UnitKey implements Built<UnitKey, UnitKeyBuilder> {
+  /// Support for serializing instances of [UnitKey].
+  static Serializer<UnitKey> get serializer => _$unitKeySerializer;
+
+  factory UnitKey({
+    @required String id,
+    @required String name,
+  }) =>
+      UnitKey._build((b) => b
+        ..id = id
+        ..name = name);
+
+  factory UnitKey._build(void Function(UnitKeyBuilder) b) = _$UnitKey;
+  UnitKey._();
+
+  /// [Unit.id].
+  String get id;
+
+  /// Name of the unit.
+  @BuiltValueField(compare: false)
+  String get name;
 }